Review fixes 1. Upgraded mattermost-redux version. 2. Moved constants.
