+++ server/gitlab/api.go
@@ -12,9 +12,9 @@ import (
 )
 
 const (
-	stateOpened  = "opened"
-	scopeAll     = "all"
-	labelDetails = true
+	stateOpened      = "opened"
+	scopeAll         = "all"
+	withLabelDetails = true
 )
 
 // NewGroupHook creates a webhook associated with a GitLab group
@@ -255,7 +255,7 @@ func (g *gitlab) GetReviews(ctx context.Context, user *UserInfo) ([]*internGitla
 
 	opened := stateOpened
 	scope := scopeAll
-	labels := labelDetails
+	labels := withLabelDetails
 
 	var result []*internGitlab.MergeRequest
 
@@ -290,7 +290,7 @@ func (g *gitlab) GetYourPrs(ctx context.Context, user *UserInfo) ([]*internGitla
 
 	opened := stateOpened
 	scope := scopeAll
-	labels := labelDetails
+	labels := withLabelDetails
 
 	var result []*internGitlab.MergeRequest
 	var resp *internGitlab.Response
@@ -332,7 +332,7 @@ func (g *gitlab) GetYourAssignments(ctx context.Context, user *UserInfo) ([]*int
 
 	opened := stateOpened
 	scope := scopeAll
-	labels := labelDetails
+	labels := withLabelDetails
 
 	var result []*internGitlab.Issue
 	var resp *internGitlab.Response
