+++ webapp/package.json
@@ -12,7 +12,7 @@
         "@primer/octicons-react": "10.1.0",
         "core-js": "3.6.5",
         "csstype": "3.0.3",
-        "mattermost-redux": "github:mattermost/mattermost-redux#b56e7e3a6f7f5de281fd3541eefec9345a8bbc89",
+        "mattermost-redux": "5.27.0",
         "prop-types": "15.7.2",
         "react": "16.13.1",
         "react-bootstrap": "1.3.0",
