+++ webapp/src/types/gitlab_items.tsx
@@ -1,18 +1,6 @@
 import {Theme} from 'mattermost-redux/types/preferences';
 import * as CSS from 'csstype';
-
-export const notificationReasons: Record<string, string> = {
-    assigned: 'You were assigned to the issue/merge request',
-    review_requested: 'You were requested to review a merge request.',
-    mentioned: 'You were specifically @mentioned in the content.',
-    build_failed: 'Gitlab build was failed.',
-    marked: 'Task is marked as done.',
-    approval_required: 'Your approval is required on this issue/merge request.',
-    unmergeable: 'This merge request can not be merged.',
-    directly_addressed: 'You were directly addressed.',
-    merge_train_removed: 'A merge train was removed.',
-    attention_required: 'Your attention is required on the issue/merge request.',
-};
+import { notificationReasons } from 'src/components/sidebar_right/gitlab_items';
 
 export interface Label {
     id: number;
